railway-deploy/requirements-railway.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy<2.0.0
pytesseract==0.3.10
openai==1.3.7
httpx==0.25.2
supabase==2.0.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
stripe==7.8.0
resend==0.6.0
